pyproject.toml
[tool.poetry]
name = "llama-index"
version = "0.1.0"
description = ""
authors = ["nehanims <neha.h.nims@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-llms-ollama = "^0.3.2"
jupyter = "^1.1.1"
llama-index-readers-file = "^0.2.1"
llama-index-vector-stores-faiss = "^0.2.1"
chromadb = "^0.5.5"
llama-index-vector-stores-chroma = "^0.2.0"
llama-index-embeddings-huggingface = "^0.3.1"
ollama = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
